Scalac: move Position/Range implicit to semanticdb
That way, we don't need to duplicate this simple code in RangePrinter.
