+++ semanticdb/scalac/library/src/main/scala/scala/meta/internal/semanticdb/scalac/SyntheticOps.scala
@@ -1,6 +1,7 @@
 package scala.meta.internal.semanticdb.scalac
 
 import scala.meta.internal.inputs._
+import scala.meta.internal.semanticdb.Implicits._
 import scala.meta.internal.semanticdb.Scala._
 import scala.meta.internal.{semanticdb => s}
